:snowman: Move config reader to libs (src/libs/config.rs, src/libs/mod.rs, src/main.rs)
