+++ src/libs/mod.rs
@@ -1,2 +1,3 @@
 pub mod api;
+pub mod config;
 pub mod version;
